+++ src/data/layout/mod.rs
@@ -1,7 +1,7 @@
 // devela/src/data/layout/mod.rs
 //
 #![doc = crate::_DOC_DATA_LAYOUT!()] // public
-#![doc = crate::_doc!(modules: crate::data; layout: array, buffer, dst, linked)]
+#![doc = crate::_doc!(modules: crate::data; layout: array, buffer, dst, linked, table)]
 #![doc = crate::_doc!(flat:"data")]
 #![doc = crate::_doc!(extends: array, collections, vec)]
 //!
@@ -15,10 +15,11 @@
 //! [`data::store`](crate::data::store).
 //!
 //! - [`Arrays`](mod@array) add dimensions and coordinate mappings.
-//! - [`Buffers`](buffer) add bounded occupancy in linear or cyclic order.
-//! - [`DST storage`](dst) supports dynamically sized representations
+//! - [`Buffers`](mod@buffer) add bounded occupancy in linear or cyclic order.
+//! - [`DST storage`](mod@dst) supports dynamically sized representations
 //!   without requiring heap allocation.
-//! - [`Linked structures`](linked) represent explicit sequential linkage.
+//! - [`Linked structures`](mod@linked) represent explicit sequential linkage.
+//! - [`Tables`](mod@table) distinguish rows, columns, and cells over tabular layouts.
 //
 
 pub mod array; // Contiguous homogeneous storage with dimensional projections
@@ -33,7 +34,7 @@ pub mod linked; // Homogeneous, sequentially accessed structures
 mod ord; // Sort
 mod queue; // Homogeneous data structures that process elements in FIFO order
 mod stack; // Homogeneous data structures that process elements in LIFO order
-// pub mod table; // TODO Tabular and heterogeneous data processing
+pub mod table; // Tabular structure over rows, columns, and cells
 
 crate::structural_mods! { // _mods, _pub_mods, _crate_internals
     _mods {
@@ -50,7 +51,7 @@ crate::structural_mods! { // _mods, _pub_mods, _crate_internals
             buffer::_all::*,
             // erased::_all::*,
             linked::_all::*,
-            // table::_all::*,
+            table::_all::*,
         };
         #[cfg_attr(not(feature = "__force_miri_dst"), cfg(not(miri)))]
         #[cfg(all(